requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.54.4
azure-data-tables==12.6.0
azure-communication-email==1.0.0
google-search-results==2.4.2
